backend/requirements.txt
annotated-types==0.7.0

antlr4-python3-runtime==4.9.3

anyio==4.6.2.post1

beautifulsoup4==4.12.3

certifi==2024.8.30

cffi==1.17.1

charset-normalizer==3.4.0

click==8.1.7

cloudpickle==3.1.0

colorama==0.4.6

decorator==4.4.2

demucs==4.0.1

dnspython==2.7.0

dora_search==0.1.12

einops==0.8.0

email_validator==2.2.0

exceptiongroup==1.2.2

fastapi==0.115.5

fastapi-cli==0.0.5

filelock==3.16.1

fsspec==2024.10.0

h11==0.14.0

httpcore==1.0.7

httptools==0.6.4

httpx==0.27.2

idna==3.10

imageio==2.36.0

imageio-ffmpeg==0.5.1

Jinja2==3.1.4

julius==0.2.7

lameenc==1.7.0

markdown-it-py==3.0.0

MarkupSafe==3.0.2

mdurl==0.1.2

moviepy==1.0.3

mpmath==1.3.0

networkx==3.2.1

numpy==2.0.2

omegaconf==2.3.0

openunmix==1.3.0

pillow==11.0.0

proglog==0.1.10

pycparser==2.22

pydantic==2.9.2

pydantic_core==2.23.4

Pygments==2.18.0

python-dotenv==1.0.1

python-multipart==0.0.17

PyYAML==6.0.2

RapidFuzz==3.10.1

requests==2.32.3

retrying==1.3.4

rich==13.9.4

shellingham==1.5.4

six==1.16.0

sniffio==1.3.1

soundfile==0.12.1

soupsieve==2.6

starlette==0.41.3

submitit==1.5.2

sympy==1.13.1

syncedlyrics==1.0.1

torch==2.5.1

torchaudio==2.5.1

tqdm==4.67.0

treetable==0.2.5

typer==0.13.1

typing_extensions==4.12.2

urllib3==2.2.3

uvicorn==0.32.0

watchfiles==0.24.0

websockets==14.1

yt-dlp==2024.11.18

